[352] client gui's
