+++ PelotonToGarmin.sln
@@ -46,6 +46,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebUI", "src\WebUI\WebUI.cs
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api", "src\Api\Api.csproj", "{63A354CC-C2D3-4860-995E-C3072EC76C46}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedUI", "src\SharedUI\SharedUI.csproj", "{5A2D34BF-6301-489C-A53C-46A2105EC923}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -88,6 +90,10 @@ Global
 		{63A354CC-C2D3-4860-995E-C3072EC76C46}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{63A354CC-C2D3-4860-995E-C3072EC76C46}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63A354CC-C2D3-4860-995E-C3072EC76C46}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A2D34BF-6301-489C-A53C-46A2105EC923}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A2D34BF-6301-489C-A53C-46A2105EC923}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A2D34BF-6301-489C-A53C-46A2105EC923}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A2D34BF-6301-489C-A53C-46A2105EC923}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
