+++ src/SharedUI/_Imports.razor
@@ -7,14 +7,12 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
+@using Microsoft.AspNetCore.Http;
 @using Common
 @using Common.Dto
 @using Common.Dto.Api
 @using Common.Dto.Peloton
 @using Common.Observe
-@using SharedUI
-@using WebUI
-@using WebUI.Shared
 @using Flurl.Http;
 @using Havit.Blazor.Components.Web
 @using Havit.Blazor.Components.Web.Bootstrap
