+++ src/WebUI/ApiClient.cs
@@ -2,31 +2,9 @@
 using Common.Dto.Api;
 using Flurl;
 using Flurl.Http;
+using SharedUI;
 
 namespace WebUI;
-
-public interface IApiClient
-{
-	Task<PelotonWorkoutsGetResponse> PelotonWorkoutsGetAsync(PelotonWorkoutsGetRequest request);
-	Task<PelotonWorkoutsGetAllResponse> PelotonWorkoutsGetAsync(PelotonWorkoutsGetAllRequest request);
-
-	Task<SettingsGetResponse> SettingsGetAsync();
-	Task<Common.App> SettingsAppPostAsync(Common.App appSettings);
-	Task<Format> SettingsFormatPostAsync(Format formatSettings);
-	Task<SettingsPelotonGetResponse> SettingsPelotonPostAsync(SettingsPelotonPostRequest pelotonSettings);
-	Task<SettingsGarminGetResponse> SettingsGarminPostAsync(SettingsGarminPostRequest garminSettings);
-
-	Task<SyncGetResponse> SyncGetAsync();
-	Task<SyncPostResponse> SyncPostAsync(SyncPostRequest syncPostRequest);
-
-	Task<SystemInfoGetResponse> SystemInfoGetAsync(SystemInfoGetRequest systemInfoGetRequest);
-
-	Task<ProgressGetResponse> GetAnnualProgressAsync();
-
-	Task<GarminAuthenticationGetResponse> GetGarminAuthenticationAsync();
-	Task<IFlurlResponse> SignInToGarminAsync();
-	Task SendGarminMfaTokenAsync(GarminAuthenticationMfaTokenPostRequest request);
-}
 
 public class ApiClient : IApiClient
 {
