+++ src/WebUI/WebUI.csproj
@@ -49,6 +49,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Api\Api.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
+    <ProjectReference Include="..\SharedUI\SharedUI.csproj" />
   </ItemGroup>
 
   <ItemGroup>
